+++ src/main/java/com/github/jlangch/venice/impl/functions/CoreFunctions.java
@@ -7260,7 +7260,7 @@ public class CoreFunctions {
 						"(supertype 5)",
 						"(supertype [1 2])",
 						"(supertype (. :java.math.BigInteger :valueOf 100))")
-					.seeAlso("type", "instance-of?")
+					.seeAlso("type", "instance-of?", "number?")
 					.build()
 		) {
 			public VncVal apply(final VncList args) {