+++ src/main/java/com/github/jlangch/venice/impl/types/VncLong.java
@@ -87,14 +87,14 @@ public class VncLong extends VncNumber {
 	
 	@Override
 	public VncKeyword getSupertype() {
-		return isWrapped() ? TYPE : VncVal.TYPE;
+		return isWrapped() ? TYPE : super.getType();
 	}
 	
 	@Override
 	public List<VncKeyword> getAllSupertypes() {
 		return isWrapped() 
 				? Arrays.asList(TYPE, VncVal.TYPE)
-				: Arrays.asList(VncVal.TYPE);
+				: super.getAllSupertypes();
 	}
 	
 	public VncLong negate() {